requirements.txt
langchain-core
langchain-huggingface
numpy<2
python-dotenv